# Conflicts:
#	api/src/main/java/com/luckygames/wmxz/gamemaster/controller/OnlineController.java
#	api/src/main/java/com/luckygames/wmxz/gamemaster/controller/PlayerCharacterController.java
#	api/src/main/java/com/luckygames/wmxz/gamemaster/model/view/base/PageQuery.java
#	api/src/main/resources/templates/index.html
#	api/src/main/resources/templates/player/forbidden.html
#	api/src/main/resources/templates/player/info.html
#	generator/src/main/resources/projects/wmxz/mybatis.xml
实时在线、批量封禁、解封、邮件、广播

+++ api/src/main/java/com/luckygames/wmxz/gamemaster/dao/handler/ForbiddenTypeHandler.java
@@ -30,7 +30,7 @@ public class ForbiddenTypeHandler implements TypeHandler<ForbiddenType> {
 
     @Override
     public void setParameter(PreparedStatement ps, int i, ForbiddenType param, JdbcType jdbcType) throws SQLException {
-        ps.setInt(i, (byte) param.code());
+        ps.setInt(i, param.code());
     }
 
 }
